Fix typos, update settings, and rename file
- Correct typos in variable names in `app/actions.tsx` and `lib/agents/researcher.tsx`.
- Add custom words to spell checker in `.vscode/settings.json`.
- Rename `history-skelton.tsx` to `history-skeleton.tsx`.
- Fix project name typo in `package.json`.
