+++ .vscode/settings.json
@@ -1,4 +1,8 @@
 {
     "editor.formatOnSave": true,
-    "editor.defaultFormatter": "esbenp.prettier-vscode"
+    "editor.defaultFormatter": "esbenp.prettier-vscode",
+    "cSpell.words": [
+        "openai",
+        "Tavily"
+    ]
 }
