+++ lib/agents/researcher.tsx
@@ -25,8 +25,8 @@ export async function researcher(
   const includeToolResponses = messages.some(message => message.role === 'tool')
   const useSubModel = useOllamaProvider && includeToolResponses
 
-  const streambleAnswer = createStreamableValue<string>('')
-  const answerSection = <AnswerSection result={streambleAnswer.value} />
+  const streamableAnswer = createStreamableValue<string>('')
+  const answerSection = <AnswerSection result={streamableAnswer.value} />
 
   const currentDate = new Date().toLocaleString()
   const result = await streamText({
@@ -51,7 +51,7 @@ export async function researcher(
     onFinish: async event => {
       finishReason = event.finishReason
       fullResponse = event.text
-      streambleAnswer.done()
+      streamableAnswer.done()
     }
   }).catch(err => {
     hasError = true
@@ -80,7 +80,7 @@ export async function researcher(
           if (useAnthropicProvider && !hasToolResult) {
             streamableText.update(fullResponse)
           } else {
-            streambleAnswer.update(fullResponse)
+            streamableAnswer.update(fullResponse)
           }
         }
         break
