+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mophic",
+  "name": "morphic",
   "version": "0.1.0",
   "private": true,
   "license": "Apache-2.0",
